Add basic command functionality
